Ensure that BP_Activity_Activity::get() pulls up comments for non-top-level items
There was previously an exception in BP_Activity_Activity::append_comments()
that would make it impossible to load activity comments for activity items
that were themselves activity comments. This changeset removes the restriction,
and adds the necessary top_level_parent_id parameter to the get_comments()
method so that the comment tree is pulled up correctly in all cases.

Fixes #4600

Props sboisvert

git-svn-id: https://buddypress.svn.wordpress.org/trunk@6410 cdf35c40-ae34-48e0-9cc9-0c9da1808c22