Change 'title' attribute for topic links in group forum loop template.
In the topic title link, bp_the_topic_title() is used twice: once
in the 'title' attribute and again for the anchor text.

For plugin devs wanting to manipulate the HTML markup for the topic
title, this would cause unintended output in the 'title' attribute.

Therefore, this commit removes bp_the_topic_title() from the 'title'
attribute for topic links.

Fixes #4770. Props chouf1.

git-svn-id: https://buddypress.svn.wordpress.org/trunk@6725 cdf35c40-ae34-48e0-9cc9-0c9da1808c22