Remove avatar location filter after handling avatar uploads
BP filters upload_dir when handling avatar uploads, to ensure that avatar files
are stored in the proper directories. This changeset ensures that the filter is
removed after use, to ensure the integrity of calls to wp_upload_dir() later on
the crop page (after the avatar upload has been handled).

Fixes #4574

git-svn-id: https://buddypress.svn.wordpress.org/trunk@6812 cdf35c40-ae34-48e0-9cc9-0c9da1808c22