Fix inaccurate active component count on BP settings screen for new installs. Fixes #4896
The members component is a mandatory component, which we always activate. The active component count only counts active components.
On new installs, the members component was not being set active, leading to the discrepancy. Saving the settings screen (without making any changes) also "fixed" the count as the save routine re-calculates the active components array.

git-svn-id: https://buddypress.svn.wordpress.org/trunk@6873 cdf35c40-ae34-48e0-9cc9-0c9da1808c22