Fixes test_bp_core_ajax_url() for new multisite unit test global manipulation
The changes introduced in r7058, which adds improvemens to go_to() for use on
multisite installations, uncovered a limitation in the test_bp_core_ajax_url()
test. Basically, it shows that bp_core_ajax_url() will always point to
admin-ajax.php of the current site, *not* the root blog. This should not be a
problem in most cases, though we should keep an eye out in the future to see
whether problems arise on mapped domains, where cross-domain AJAX calls may be
affected.

See #4761

git-svn-id: https://buddypress.svn.wordpress.org/trunk@7059 cdf35c40-ae34-48e0-9cc9-0c9da1808c22