Improves the algorithm for filtering by role in BP_Group_Member_Query
Filtering by group role is tricky, because group role data is declared
explicitly in the database only for 'admin', 'mod', and 'banned' - 'member'
status is implicit, when these other three values are 0. The improved algorithm
assembles a role whitelist in cases where the list of roles being filtered does
not include 'member', and a role blacklist in the case where the list does
contain 'member'.

git-svn-id: https://buddypress.svn.wordpress.org/trunk@7144 cdf35c40-ae34-48e0-9cc9-0c9da1808c22