Introduces bp_activity_comment_permalink() and implements in comment template
ID anchors have always been available for the purpose of linking directly to
specific activity comments. But these links have never been available directly
in the interface: the timestamp of specific comments lead to the parent thread,
instead of to the specific comment within that thread. This changeset will make
it easier to find the permalink of specific comments.

Fixes #4741

git-svn-id: https://buddypress.svn.wordpress.org/trunk@7168 cdf35c40-ae34-48e0-9cc9-0c9da1808c22