Refines the way that BP_ROOT_BLOG is calculated on multisite installations
BP was originally a 'network: true' plugin, making it possible to assume blog 1
as the default value of BP_ROOT_BLOG on a multisite installation, unless
otherwise declared. When the network-only restriction was lifted, it became
possible to activate BP only on a secondary site in a network, in which case
the default value 1 of BP_ROOT_BLOG no longer applies.

This changeset reworks the logic used to calculate BP_ROOT_BLOG, starting with
the assumption that it is the current blog, and only changing it to the
current site's blog_id (blog 1, in non-multi-network setups) when the plugin
is indeed network-activated.

This fixes a number of problems, notably one where it was impossible to set up
the bp-pages associations, because the Pages dropdowns were populated from blog
1 when BP was active only on a secondary blog of a network.

Fixes #4917

Props r-a-y

git-svn-id: https://buddypress.svn.wordpress.org/trunk@7192 cdf35c40-ae34-48e0-9cc9-0c9da1808c22