Only show the "Edit Member > User Capabilities" toolbar item if the settings component is enabled.
The "User Capabilities" page requires the Settings component to be active
in order for the page to display.

Since, by default, BuddyPress does not have the Settings component active,
if a super admin attempts to click on this link, it will 404.

Adding a check for the Settings component fixes this.

git-svn-id: https://buddypress.svn.wordpress.org/trunk@7226 cdf35c40-ae34-48e0-9cc9-0c9da1808c22