Scroll to top when clicking on bottom pagination links in BP loops.
If a user clicks on a pagination link located at the bottom of any BP loop
(members, groups), the loop will refresh with AJAX, but will not scroll to
the top of the page.  This isn't so great for UX.

This commit adds the ability to scroll to top when a bottom-paginated
link is clicked.

In order for this to work, an additional parameter ('caller') was added to
the bp_filter_request() JS function to determine where the link was called
from.

Fixes #5206.

git-svn-id: https://buddypress.svn.wordpress.org/trunk@7470 cdf35c40-ae34-48e0-9cc9-0c9da1808c22