Refactor get_group_extras logic in group loops
The previous logic made it so that is_pending, is_invited, and is_member
values were only set in the case where users had a membership for the group
in the database. This could result in PHP warnings in certain cases, and in
all cases results in inconsistent data structures between member and non-member
groups within group loops.

Fixes #5269

git-svn-id: https://buddypress.svn.wordpress.org/trunk@7616 cdf35c40-ae34-48e0-9cc9-0c9da1808c22