Unit Tests: switch to using `develop.svn.wordpress.org` for the travis-ci test environment instead of `unit-tests.svn.wordpress.org`.
`unit-tests.svn` is a deprecated repo and `develop.svn`'s test suite has diverged from it to the point where there are a couple of important differences in the test environment that affect how BuddyPress unit tests are run on travis-ci; the defaults for `error_reporting` and handling of functions that have been marked as deprecated.

Recent commits such as r8015 have been successful in adjusting the `error_reporting` level but our tests that have been marked as `@expectedDeprecated` are currently not handled correctly on travis-ci and on any local testing environments that use `unit-tests.svn` along with a development `error_reporting` level.

A consequence of switching the travis-ci script to `develop.svn` is that, as the WP core-to-develop SVN restructing happened between WP 3.6 and 3.7, SVN versions of WordPress prior to 3.7 don't have the seperate `src` and `tests` folders which the updated travis-ci script relies on. The tradeoff is that we'll have increased confidence in our travis-ci environment going forwards.

git-svn-id: https://buddypress.svn.wordpress.org/trunk@8017 cdf35c40-ae34-48e0-9cc9-0c9da1808c22