Groups: When fetching a new group object, default 'populate_extras' to false.
In r7934, the 'populate_extras' parameter was set to true.  This
parameter queries additional info about the group including group
admins, moderators and member status.

There are very few instances where we need this extra information.  We
only need this data when on a single group or in a group loop.

Therefore, this commit sets the 'populate_extras' parameter to false
by default and sets it to true when setting up the current group
object.

Fixes #5411

git-svn-id: https://buddypress.svn.wordpress.org/trunk@8056 cdf35c40-ae34-48e0-9cc9-0c9da1808c22