Introduce BuddyPress Tools panel
Modeled on the bbPress panel, the BuddyPress Tools panel contains a number of
tools for repairing BuddyPress data. In this first iteration, the following
actions are supported:

- Recalculate total_friend_count for each user
- Recalculate total_group_count for each user
- Recalculate total site members

On non-Multisite (or when BP is only activated on a single site), the menu
appears as Tools > BuddyPress. When BP is network-activated, we create a stub
Tools menu on the Network Admin to account for the fact that one is not
normally there. If WP ever introduces a menu like this, we can use it instead.

Fixes #5296

Props cmmarslender, boonebgorges, imath

git-svn-id: https://buddypress.svn.wordpress.org/trunk@8117 cdf35c40-ae34-48e0-9cc9-0c9da1808c22