Don't trim whitespace from $meta_value param in _delete_meta() functions
The $meta_value parameter of BP's delete_meta() functions allows developers to
limit deletion to rows that match both the $meta_key *and* the specified
$meta_value. However, most of BP's delete_meta() functions have always had a
quirk whereby the $meta_value was run through trim() before matching against
the values in the database. This had the unacceptable side effect that passing
a $meta_value of '  foo  ' would delete rows with the value 'foo', but *not*
the value '  foo  '!

This changeset reverses the behavior, and modifies the relevant unit tests.

See #5399

git-svn-id: https://buddypress.svn.wordpress.org/trunk@8130 cdf35c40-ae34-48e0-9cc9-0c9da1808c22