Display a Manage Signups submenu to Users menu in wp-admin
Signups management is simulating a new view in the wp-admin users page. A Pending link is available after the different WordPress views to filter the users by roles. It appears the Signups screen can be a bit difficult to find for site or network administrators. As a result, it is best to display the Manage Signups submenu instead of removing it.

See #5374

props johnjamesjacoby, boonebgorges

git-svn-id: https://buddypress.svn.wordpress.org/trunk@8136 cdf35c40-ae34-48e0-9cc9-0c9da1808c22