Don't register the bp-themes themes directory if bp-default is found elsewhere in the installation
This will be part of the migration path away from the packaged version of
bp-default. bp-default will be added to the wordpress.org theme repository with
a higher version number, prompting an upgrade that will result in an
installation into the standard WP themes directory. Once that happens, most
installations will no longer be reliant on the version of bp-default packaged
with BuddyPress.

See #5212

Props DJPaul

git-svn-id: https://buddypress.svn.wordpress.org/trunk@8158 cdf35c40-ae34-48e0-9cc9-0c9da1808c22