Activity dashboard: Check if activity item can be replied to.
If activity commenting is enabled for blog posts, in the activity
dashboard, we have to check whether or not the activity item can be
replied to.

This commit adds the class method - BP_Activity_List_Table::can_comment() -
to check the current activity item in the activity list table.

See #5130.

git-svn-id: https://buddypress.svn.wordpress.org/trunk@8194 cdf35c40-ae34-48e0-9cc9-0c9da1808c22