Only load the activity just posted when on a user activity page
HeartBeat activity check feature introduced a "since" argument to load the activity that are newest. In a user activity page, this feature is not used but a timestamp is created using a "0" value causing all user activities to be reloaded. Checking if the value is empty before creating the timestamp will make sure the activity just posted will be the only one loaded into user stream.

props boonebgorges, imath

See #5603

git-svn-id: https://buddypress.svn.wordpress.org/trunk@8341 cdf35c40-ae34-48e0-9cc9-0c9da1808c22