Disable COUNT queries by default in bp_activity_get() stack
The COUNT query performed to get the total activity count has proven to be a
performance bottleneck on large sites. Moreover, the way that activity items
are displayed on the front end (using the Load More link) means that the total
count is not actually used by BuddyPress in most cases. To reduce query
overhead, we introduce a 'count_total' parameter to the bp_activity_get()
stack, and set it to false by default.

For backward compatibility, a few additional changes are introduced:

- In the activity-loop.php template, the <noscript> pagination markup is
  removed. In its place, the Load More link is refactored so that it loads
  the next available page of activity items.
- The mechanism used to determine whether there are more activity items to show
  (bp_activity_has_more_items()) has been refined; when no COUNT query takes
  place, we query for the $per_page value + 1 to infer whether more items are
  available.

Fixes #5629

Props boonebgorges, r-a-y

git-svn-id: https://buddypress.svn.wordpress.org/trunk@8491 cdf35c40-ae34-48e0-9cc9-0c9da1808c22